requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.5.0
python-telegram-bot==20.6
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.27.0
openai==1.3.8